Redesign product management backend and frontend

+++ database/migrations/2025_04_27_071442_create_products_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained()->cascadeOnDelete()->cascadeOnUpdate();
             $table->string('name');
-            $table->float('price', 2);
+            $table->decimal('price', 12, 2);
             $table->text('description')->nullable();
-            $table->integer('quantity');
+            $table->unsignedInteger('quantity');
             $table->timestamps();
         });
     }
